+++ src/graysvr/CCharUse.cpp
@@ -182,7 +182,7 @@ bool CChar::Use_Kindling(CItem *pItem)
 
 	pItem->SetID(ITEMID_CAMPFIRE);
 	pItem->SetAttr(ATTR_MOVE_NEVER|ATTR_DECAY);
-	pItem->SetTimeout((4 + pItem->GetAmount()) * 60 * TICK_PER_SEC);
+	pItem->SetTimeout((4 + static_cast<INT64>(pItem->GetAmount())) * 60 * TICK_PER_SEC);
 	pItem->SetAmount(1);	// all kindling is set to one fire
 	pItem->m_itLight.m_pattern = LIGHT_LARGEST;
 	pItem->Update();
